fix(fsdp): default the cond input axis to pass-through
The base default cond: "default" claimed to mirror the wan/ltx rollout
boundary but does not survive contact with paired dumps across three
families (baseline side = actual rollout entry dtypes):

  family    latents  timestep  text cond (encoder_hidden_states)
  wan2.2    bf16     fp32      fp32 -- rollout does NOT cast
  ltx-2.3   bf16     fp32      bf16 -- already target dtype, cast is a no-op
  sd3.5     fp16     fp32      fp32 -- rollout does NOT cast

Wan rollout only casts image_embeds at the boundary (I2V); the UMT5/T5
text output enters the DiT in fp32 and the first context linear consumes
it with reduced-precision weights (autocast re-quantizes at the matmul).
Casting cond early therefore diverges from rollout on wan/sd3 (the old
wan runs carried a rel ~1.7e-3 rounding scar at the condition_embedder
entry from exactly this cast) and is a no-op on ltx. cond: None is exact
for all three. The only scenario cond: "default" ever guarded --
trajectory serialization upcasting cond to fp32 -- has not been observed
on the cond path in any family.

+++ tests/fast/backends/fsdp_utils/test_input_dtype_policy.py
@@ -32,7 +32,9 @@ def test_default_policy_matches_rollout_boundary():
     )
     assert out_latents.dtype == torch.bfloat16
     assert out_timesteps.dtype == torch.float32
-    assert out_pos["context"].dtype == torch.bfloat16
+    # cond passes through: dump-verified on wan2.2/sd3.5 (rollout feeds fp32
+    # text embeds straight into the first context linear) and a no-op on ltx.
+    assert out_pos["context"].dtype == torch.float32
     assert out_pos["context_mask"].dtype == torch.int64
     assert out_neg is None and out_joint is None
